Fixed a nasty waterfall glitch on Linux
git-svn-id: https://subversion.assembla.com/svn/sdrsharp/trunk@630 d3cde5d6-82b8-4669-8b62-43054d6781d8

+++ PanView/Waterfall.cs
@@ -471,7 +471,7 @@ namespace SDRSharp.PanView
             }
             else
             {
-                ptr = (int*) bits.Scan0 + 4 * bits.Width * (bits.Height - 1) + AxisMargin;
+                ptr = (int*) ((long) bits.Scan0 + 4 * bits.Width * (bits.Height - 1)) + AxisMargin;
             }
             for (var i = 0; i < _spectrum.Length; i++)
             {
